added more socials
