+++ src/app/events/moresocials/page.tsx
@@ -1,0 +1,11 @@
+import MoreSocials from "@/components/events/moresocials/moreSocials";
+
+const Page = () => {
+  return (
+    <div>
+      <MoreSocials />
+    </div>
+  );
+};
+
+export default Page;
